Make the pip "manager" works on windows.
Tested on Windows 10 under git bash.
