+++ meta_package_manager/managers/pip.py
@@ -32,12 +32,12 @@ from boltons.cacheutils import cachedproperty
 import simplejson as json
 
 from ..base import PackageManager
-from ..platform import LINUX, MACOS
+from ..platform import LINUX, MACOS, WINDOWS
 
 
 class Pip(PackageManager):
 
-    platforms = frozenset([MACOS, LINUX])
+    platforms = frozenset([MACOS, LINUX, WINDOWS])
 
     requirement = '>= 9.0.0'
