+++ meta_package_manager/tests/test_managers.py
@@ -152,4 +152,4 @@ class TestManagerPlatform(unittest.TestCase):
     @unless_windows()
     def test_windows(self):
         supported_managers = {m.id for m in pool().values() if m.supported}
-        self.assertSetEqual(supported_managers, set())
+        self.assertSetEqual(supported_managers, set(['pip2', 'pip3']))
